requirements.txt
# Core Web Framework
streamlit>=1.28.0

# Git Operations
gitpython>=3.1.40
pygithub>=1.59.1

# Database
neo4j>=5.14.0
redis>=5.0.0

# AI/LLM Integration
openai>=1.3.0
langchain>=0.0.350
langgraph>=0.0.20
langchain-community>=0.0.10

# Data Processing
pydantic>=2.5.0
requests>=2.31.0

# Security & Encryption
cryptography>=41.0.0

# Code Analysis Tools
flake8>=6.1.0
pylint>=3.0.0
black>=23.11.0
mypy>=1.7.0
isort>=5.12.0

# CLI Framework
click>=8.1.7
typer>=0.9.0

# Logging
loguru>=0.7.2

# Testing Framework
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Development Tools
pre-commit>=3.5.0

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=1.3.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
jinja2>=3.1.0
matplotlib>=3.8.0
plotly>=5.17.0
pandas>=2.1.0

# Configuration File Parsing
tomli>=2.0.1  # For pyproject.toml parsing
